🪛 feat: improve build config
