+++ package.json
@@ -37,11 +37,9 @@
     "docs": "cd website && npm run start",
     "build:docs": "cd website && npm run build",
     "dev": "vite",
-    "types": "tsc --project tsconfig.types.json",
-    "build": "node scripts/clearBuildFolder.js && pnpm run build:only && pnpm run types",
     "build:legacy::only": "vite build",
     "build:legacy:watch": "vite build --watch",
-    "build:only": "rollup -c",
+    "build": "rollup -c",
     "build:watch": "rollup -c -w",
     "prepublishOnly": "pnpm run build",
     "publish:alpha": "cross-env BUILD_TAG=alpha npm publish --tag alpha",
@@ -84,6 +82,8 @@
     "@emotion/react": "^11.9.0",
     "@emotion/styled": "^11.8.1",
     "@jest/types": "27",
+    "@rollup/plugin-commonjs": "^23.0.2",
+    "@rollup/plugin-node-resolve": "^15.0.1",
     "@stitches/react": "^1.2.8",
     "@swc/core": "^1.2.154",
     "@swc/jest": "^0.2.20",
@@ -126,9 +126,9 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-is": "^18.2.0",
-    "rollup": "^2.77.2",
-    "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-typescript2": "^0.31.2",
+    "rollup": "^3.2.3",
+    "rollup-plugin-dts": "^5.0.0",
+    "rollup-plugin-esbuild": "^4.10.1",
     "sass": "^1.50.0",
     "styled-components": "^5.3.5",
     "tailwindcss": "^3.1.2",
