--- tsconfig.types.json
@@ -1,11 +1,0 @@
-{
-  "compilerOptions": {
-    "jsx": "react-jsx",
-    "incremental": true,
-    "declaration": true,
-    "emitDeclarationOnly": true,
-    "esModuleInterop": true,
-    "outDir": "dist"
-  },
-  "include": ["src/index.ts"],
-}
